# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "嘉阅圈私域社群平台后端服务，提供内容社区、互动打卡、小组交流、私信通知、知识课程等核心能力。\n统一响应格式：{\"success\":bool,\"data\":any,\"message\":string,\"error\":{\"code\":string,\"message\":string}}",
        "title": "嘉阅圈后端 API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3100",
    "basePath": "/api",
    "paths": {
...
        "/users/profile": {
            "put": {
                "description": "需登录。仅更新请求体中实际传入的字段，支持昵称、头像、简介、性别与生日；头像需通过持久化 URL 归属校验，性别取值 UNKNOWN/MALE/FEMALE，生日支持 RFC3339、ISO 日期或秒级时间戳。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "更新当前用户资料",
                "parameters": [
                    {
                        "description": "更新资料请求",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数格式错误或性别/生日/头像取值无效",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "未登录或登录已过期",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "description": "公开接口。返回指定用户的公开主页，包含昵称、头像、简介、性别、连续打卡天数（按业务日期计算）、关注数与粉丝数。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "获取用户公开资料",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "无效的用户 ID",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "用户不存在",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}/follow": {
            "post": {
                "description": "需登录。关注指定用户，校验目标用户存在性、防止自我关注与重复关注，并同步更新双方关注/粉丝计数。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "关注用户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "目标用户 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "无效的用户 ID 或不能关注自己",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "未登录或登录已过期",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "409": {
                        "description": "已经关注该用户",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "需登录。取消关注指定用户，校验关注关系是否存在，并同步更新双方关注/粉丝计数。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "取消关注用户",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "目标用户 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "无效的用户 ID 或不能关注自己",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "未登录或登录已过期",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "409": {
                        "description": "尚未关注该用户",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}/followers": {
            "get": {
                "description": "公开接口。分页查询指定用户的粉丝列表，page 默认 1，pageSize 默认 20、上限 100。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "查询用户粉丝列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码（默认 1）",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量（默认 20，上限 100）",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "无效的用户 ID",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/users/{id}/following": {
            "get": {
                "description": "公开接口。分页查询指定用户的关注列表，page 默认 1，pageSize 默认 20、上限 100。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "查询用户关注列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "页码（默认 1）",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量（默认 20，上限 100）",
                        "name": "pageSize",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "无效的用户 ID",
                        "schema": {
                            "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "yjedu-reading-club-server_internal_pkg_response.ApiError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "yjedu-reading-club-server_internal_pkg_response.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/yjedu-reading-club-server_internal_pkg_response.ApiError"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "输入 \"Bearer \u003ctoken\u003e\"，token 通过 /api/auth/login/phone 或 /api/auth/weapp/login 获取。",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}